feat-bo: replace useAuth hook by Session context and session hook
