+++ backoffice/src/hooks/useSession.ts
@@ -1,0 +1,6 @@
+import { useContext } from "react";
+import { SessionContext } from "src/contexts/SessionContext";
+
+const useSession = () => useContext(SessionContext);
+
+export default useSession
